third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/content-reprojection-reattach-expected.txt

CONSOLE ERROR: line 9: Uncaught TypeError: host.createShadowRoot is not a function
